+++ src/kernel/input/known_functions.h
@@ -1,8 +1,8 @@
 #ifndef KNOWN_FUNCTIONS_H
 #define KNOWN_FUNCTIONS_H
 
-double f0 (const int t, const int x);
-double f1 (const int t, const int x);
+double f0 (const double t, const double x);
+double f1 (const double t, const double x);
 
 double p_wave_deriv (const double z);
